+++ app/src/main/java/me/saket/dank/urlparser/UrlParserConfig.java
@@ -8,6 +8,8 @@ import javax.inject.Singleton;
 @Singleton
 public class UrlParserConfig {
 
+  private static final String IMGUR_LOW_QUALITY_MAXRES = "1280";
+
   /**
    * /r/$subreddit.
    */
@@ -72,11 +74,14 @@ public class UrlParserConfig {
   private static final Pattern DEFAULT_IMGUR_ALBUM_PATTERN = Pattern.compile("/(?:gallery)?(?:a)?(?:t/\\w*)?/(\\w*).*");
 
   /**
-   * Extracts Imgur preview suffix and file extension
+   * Extracts Imgur image id (1) and extension (2) if exists
    * <p>
-   * _d.jpg
+   * /TwxFuFG_d.jpg
+   * /TwxFuFG.jpg
+   * /TwxFuFG.png
+   * /TwxFuFG
    */
-  private static final Pattern IMGUR_PREVIEW_SUFFIX_PATTERN = Pattern.compile("_d(\\.\\w+)$");
+  private static final Pattern IMGUR_IMAGE_ID_EXT_PATTERN = Pattern.compile("^.*/([A-Za-z0-9]+)(?:_d)?(?:\\.(\\w+))?$");
 
   @Inject
   public UrlParserConfig() {
@@ -134,7 +139,11 @@ public class UrlParserConfig {
     return DEFAULT_IMGUR_ALBUM_PATTERN;
   }
 
-  Pattern imgurPreviewExtPattern() {
-    return IMGUR_PREVIEW_SUFFIX_PATTERN;
+  public Pattern imgurIdExtPattern() {
+    return IMGUR_IMAGE_ID_EXT_PATTERN;
+  }
+
+  public String imgurLowQualityMaxres() {
+    return IMGUR_LOW_QUALITY_MAXRES;
   }
 }